Alignments with branch next.
