+++ SublimeLinter.sublime-settings
@@ -21,9 +21,6 @@
     // to an icon in the theme (they usually include "warning" and "error")
     "gutter_theme": "none",
 
-    // Lints all open views on startup. Recommended only for faster systems.
-    "lint_all_views_on_startup": false,
-
     // Lint Modes determine when the linter is run
     // background: asynchronously on every change
     // load_save: when a file is opened and every time it's saved
