+++ lint/settings.py
@@ -173,7 +173,8 @@ class Settings:
             self.update_gutter_icons()
 
         Linter.reload()  # always reload
-        sublime.active_window().run_command("sublime_linter_lint", {"lint_all_views": True})
+        from ..sublime_linter import SublimeLinter
+        SublimeLinter.lint_all_views()
 
         self.changeset.clear()
